Add the javascript device.nfc variable
Adds the javascript variable "nfc" to the "device" object to allow other
code to check if the NFC plugin was successfully initalized and if the
phone supports NFC.
If the phone has NFC support, then the variable device.nfc will be true,
false otherwise.

Note: a real phone support check has been only added to Android.